allow square brackets in regular expressions
Until now the square brackets were taken as the opening brackets of selectors.
Create a state for regular expressions to get a new matching scope. This also
needs fixing of the p_condition_regex() function, which called an undefined
class ConditionRegex. This was not hit until now because the rule never
matched, until now regular expressions were matched as identifiers. From now on
slashes are not allowed in identifiers anymore.